+++ apps/api/src/app/portfolio/calculator/roai/portfolio-calculator-cash.spec.ts
@@ -231,6 +231,7 @@ describe('PortfolioCalculator', () => {
        */
       expect(position).toMatchObject<TimelinePosition>({
         activitiesCount: 2,
+        annualizedDividendYield: 0,
         averagePrice: new Big(1),
         currency: 'USD',
         dataSource: DataSource.YAHOO,