+++ apps/api/src/app/portfolio/portfolio.service.ts
@@ -1862,8 +1862,11 @@ export class PortfolioService {
       }
     }
 
-    const { currentValueInBaseCurrency, totalInvestment } =
-      await portfolioCalculator.getSnapshot();
+    const {
+      annualizedDividendYield,
+      currentValueInBaseCurrency,
+      totalInvestment
+    } = await portfolioCalculator.getSnapshot();
 
     const { performance } = await this.getPerformance({
       impersonationId,
@@ -1961,6 +1964,7 @@ export class PortfolioService {
       })?.toNumber();
 
     return {
+      annualizedDividendYield,
       annualizedPerformancePercent,
       annualizedPerformancePercentWithCurrencyEffect,
       cash,